[add] no breaking changes on autocompleter
