+++ redisearch/autocomplete_test.go
@@ -1,11 +1,23 @@
 package redisearch_test
 
 import (
+	"fmt"
 	"github.com/RediSearch/redisearch-go/redisearch"
 	"github.com/garyburd/redigo/redis"
+	"github.com/stretchr/testify/assert"
+	"os"
 	"reflect"
 	"testing"
 )
+
+func createAutocompleter(dictName string) *redisearch.Autocompleter {
+	value, exists := os.LookupEnv("REDISEARCH_TEST_HOST")
+	host := "localhost:6379"
+	if exists && value != "" {
+		host = value
+	}
+	return redisearch.NewAutocompleter(host, dictName)
+}
 
 func TestAutocompleter_Serialize(t *testing.T) {
 	fuzzy := redisearch.DefaultSuggestOptions
@@ -41,7 +53,7 @@ func TestAutocompleter_Serialize(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			a := redisearch.NewAutocompleter(tt.fields.name)
+			a := redisearch.NewAutocompleterFromPool(nil,tt.fields.name)
 			got, got1 := a.Serialize(tt.args.prefix, tt.args.opts)
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("Serialize() got = %v, want %v", got, tt.want)
@@ -50,5 +62,40 @@ func TestAutocompleter_Serialize(t *testing.T) {
 				t.Errorf("Serialize() got1 = %v, want %v", got1, tt.want1)
 			}
 		})
+	}
+}
+
+func TestSuggest(t *testing.T) {
+	a := createAutocompleter("testing")
+
+	// Add Terms to the Autocompleter
+	terms := make([]redisearch.Suggestion, 10)
+	for i := 0; i < 10; i++ {
+		terms[i] = redisearch.Suggestion{Term: fmt.Sprintf("foo %d", i),
+			Score: 1.0, Payload: fmt.Sprintf("bar %d", i)}
+	}
+	err := a.AddTerms(terms...)
+	assert.Nil(t, err)
+	suglen, err := a.Length()
+	assert.Nil(t, err)
+	assert.Equal(t, int64(10), suglen)
+	// Retrieve Terms From Autocompleter - Without Payloads / Scores
+	suggestions, err := a.SuggestOpts("f", redisearch.SuggestOptions{Num: 10})
+	assert.Nil(t, err)
+	assert.Equal(t, 10, len(suggestions))
+	for _, suggestion := range suggestions {
+		assert.Contains(t, suggestion.Term, "foo")
+		assert.Equal(t, suggestion.Payload, "")
+		assert.Zero(t, suggestion.Score)
+	}
+
+	// Retrieve Terms From Autocompleter - With Payloads & Scores
+	suggestions, err = a.SuggestOpts("f", redisearch.SuggestOptions{Num: 10, WithScores: true, WithPayloads: true})
+	assert.Nil(t, err)
+	assert.Equal(t, 10, len(suggestions))
+	for _, suggestion := range suggestions {
+		assert.Contains(t, suggestion.Term, "foo")
+		assert.Contains(t, suggestion.Payload, "bar")
+		assert.NotZero(t, suggestion.Score)
 	}
 }
