+++ redisearch/client.go
@@ -422,119 +422,3 @@ func (i *Client) Info() (*IndexInfo, error) {
 
 	return &ret, nil
 }
-
-// Delete deletes the Autocompleter key for this AC
-func (i *Client) AutocompleterDelete(a *Autocompleter) error {
-	conn := i.pool.Get()
-	defer conn.Close()
-
-	_, err := conn.Do("DEL", a.name)
-	return err
-}
-
-// AddTerms pushes new term suggestions to the index
-func (c *Client) AutocompleterAddTerms(a *Autocompleter, terms ...Suggestion) error {
-	conn := c.pool.Get()
-	defer conn.Close()
-
-	i := 0
-	for _, term := range terms {
-
-		args := redis.Args{a.name, term.Term, term.Score}
-		if term.Payload != "" {
-			args = append(args, "PAYLOAD", term.Payload)
-		}
-
-		if err := conn.Send("FT.SUGADD", args...); err != nil {
-			return err
-		}
-		i++
-	}
-	if err := conn.Flush(); err != nil {
-		return err
-	}
-	for i > 0 {
-		if _, err := conn.Receive(); err != nil {
-			return err
-		}
-		i--
-	}
-	return nil
-}
-
-// AddTerms pushes new term suggestions to the index
-func (c *Client) AutocompleterDeleteTerms(a *Autocompleter, terms ...Suggestion) error {
-	conn := c.pool.Get()
-	defer conn.Close()
-
-	i := 0
-	for _, term := range terms {
-
-		args := redis.Args{a.name, term.Term}
-		if err := conn.Send("FT.SUGDEL", args...); err != nil {
-			return err
-		}
-		i++
-	}
-	if err := conn.Flush(); err != nil {
-		return err
-	}
-	for i > 0 {
-		if _, err := conn.Receive(); err != nil {
-			return err
-		}
-		i--
-	}
-	return nil
-}
-
-// AddTerms pushes new term suggestions to the index
-func (c *Client) AutocompleterLength(a *Autocompleter) (len int64, err error) {
-	conn := c.pool.Get()
-	defer conn.Close()
-	len, err = redis.Int64(conn.Do("FT.SUGLEN", a.name))
-	return
-}
-
-// Suggest gets completion suggestions from the Autocompleter dictionary to the given prefix.
-// If fuzzy is set, we also complete for prefixes that are in 1 Levenshten distance from the
-// given prefix
-//
-// Deprecated: Please use SuggestOpts() instead
-func (c *Client) AutocompleterSuggest(a *Autocompleter, prefix string, num int, fuzzy bool) (ret []Suggestion, err error) {
-	conn := c.pool.Get()
-	defer conn.Close()
-
-	seropts := DefaultSuggestOptions
-	seropts.Num = num
-	seropts.Fuzzy = fuzzy
-	args, inc := a.Serialize(prefix, seropts)
-
-	vals, err := redis.Strings(conn.Do("FT.SUGGET", args...))
-	if err != nil {
-		return nil, err
-	}
-
-	ret = ProcessSugGetVals(vals, inc, true, false)
-
-	return
-}
-
-// SuggestOpts gets completion suggestions from the Autocompleter dictionary to the given prefix.
-// SuggestOptions are passed allowing you specify if the returned values contain a payload, and scores.
-// If SuggestOptions.Fuzzy is set, we also complete for prefixes that are in 1 Levenshten distance from the
-// given prefix
-func (c *Client) AutocompleterSuggestOpts(a *Autocompleter, prefix string, opts SuggestOptions) (ret []Suggestion, err error) {
-	conn := c.pool.Get()
-	defer conn.Close()
-
-	args, inc := a.Serialize(prefix, opts)
-	vals, err := redis.Strings(conn.Do("FT.SUGGET", args...))
-	if err != nil {
-		return nil, err
-	}
-
-	ret = ProcessSugGetVals(vals, inc, opts.WithScores, opts.WithPayloads)
-
-	return
-}
